feat(audio): Add audio file upload and query functionality
- Added audio file upload API, supporting file storage and metadata recording
- Implemented audio file pagination query API, with conditional filtering support
- Added audio file list query API, returning file view object list
- Refined AudioFile entity fields: changed status and delete flag types to Integer
- Added AudioFileQueryRequest DTO extending PageRequest to support pagination
- Created AudioFileVO view object, containing full audio file information with object conversion support
- Implemented getQueryWrapper method for dynamic query construction
- Upgraded project JDK version to 17
- Removed redundant Redis dependency
- Added serialVersionUID to multiple DTOs and entities
- Optimized scope of shift constants in SnowflakeIdGenerator
- Fixed byte assignment and comparison logic in UUID utility class
- Updated frontend upload component code formatting and function signatures
- Added frontend ESLint + Prettier and TypeScript configuration dependencies
- Added SELECT type in BusinessType for logging support
- Optimized FileUtils UTF-8 encoding by using StandardCharsets reference
- Removed unused author comments in DateUtils
- Removed redundant UTF8 constant definition in FileUtils

+++ server/core/src/main/java/com/carbon/audio/model/entity/AudioFile.java
@@ -59,7 +59,7 @@ public class AudioFile implements Serializable {
     /**
      * 文件状态（0 - 待处理、1 - 处理中、2 - 处理成功、3 - 处理失败）
      */
-    private Long status;
+    private Integer status;
 
     /**
      * 保留字段1
@@ -79,7 +79,7 @@ public class AudioFile implements Serializable {
     /**
      * 是否删除
      */
-    private Long isDelete;
+    private Integer isDelete;
     /**
      * 创建时间
      */
